Add default tests

+++ python/xecs/_internal/float32.py
@@ -3,5 +3,5 @@ from typing import cast
 from xecs.xecs import Float32
 
 
-def float32(*, default: int) -> Float32:
+def float32(*, default: float) -> Float32:
     return cast(Float32, default)
